requirements.txt
asgiref==3.5.2
Django==4.1.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
Pillow==9.3.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
